lsp,ui: do not attach lsp clients
previous to this commit the calltree buffer would attach active lsp
clients.

this was for an initial idea which never panned out, but also causes
issues as the LSP think the calltree buffer is a valid source file, it
is not.

additionally, this pr fixes a few bad variable references in ui.lua
w/r/t config.

closes: #2

Signed-off-by: ldelossa <louis.delos@gmail.com>