tree: unique keys
to get around #10 each node in the tree now has a unique key.

marshaling the tree is updated to cache the node at all given lines every
time write_tree is called.

marshaling a line is greatly simplified and just looks at the cached now
to understand which node is being selected.

closes #10

Signed-off-by: ldelossa <louis.delos@gmail.com>