ui: declarative layout
this commit adds a way for the UI to request opening a window in a
declarative fashion.

the ui.window module is now updated to compute the current and desired
layouts and then perform the actions to get neovim to this desired
layout.

the hope is for this to aid easily adding new windows to the calltree
ui.

symboltree is stubbed out (not yet implemented) to provide a proof of
concept.

Signed-off-by: ldelossa <louis.delos@gmail.com>