ui: use lsp detail if file is not in current workspace
closes #16

this commit changes the logic of marshaling a node to a UI line.

if we determine the file a symbol belongs too does not exist in the
current workspace we now use the "detail" field from the symbol if
possible.

if the "detail" field is nil we fall back to just providing the full
path to the file.

Signed-off-by: ldelossa <louis.delos@gmail.com>