ui: display no file info for external packages
closes #16

this commit takes a slightly modified approach from vscode's when files from external
modules/packages are displayed in the calltree.

if the package is from an external module and the LSP returns no
"detail" field we then display no file info at all.

if the user still wants information about where the symbol is coming
from they can use the "CTDetails" command when their cursor is on the
symbol.

Signed-off-by: ldelossa <louis.delos@gmail.com>