perf: async symbol gathering
LSPs may return different results depending on the method being
requested.

for instance gopls will return no receiver name when a call hierarchy
request is made, but will when a workspace symbol request is made.

to work around this we make a workspace symbol request for each call
hierarchy item returned when a calltree is invoked.

this proved to be very slow when done sync so this commit makes it
async using lua coroutines.

Signed-off-by: ldelossa <louis.delos@gmail.com>