ui: unified panel
this commit adds the ability to treat all calltree ui elements as a
unified panel similar to other IDEs.

the unified panel consists of a persistent symbol and call tree and both
UIs will stick together when the unified panel is toggled open or
closed.

this is an opt in feature and using each component separately is still
possible.

Signed-off-by: ldelossa <louis.delos@gmail.com>