ui: open directly to calltree ui elements
this commit reworks the CTOpen and STOpen commands.

now calling these commands will open the respective UI and jump
directly into it.

calling the same command from inside the calltree ui window will jump
back to the window which invoked it.

this is very useful when "jumping over" split windows such that the
symboltree does not unintentionally update with the outline of a buffer
between the intended buffer and the symbol outline window.

Signed-off-by: ldelossa <louis.delos@gmail.com>