workflow: add smart closing of UI elements
this commit adds a keymap to all calltree ui buffers for closing the
element and jumping back to the window that invoked it (if it still
exists).

this is useful for workflows which do not use a unified panel but also
like to quickly hide the UI elements after browsing the call/symbol
tree.

Signed-off-by: ldelossa <louis.delos@gmail.com>