usability: unified panel by default
this commit reworks the idea of an optional unified panel as the idea
was getting messy.

now, the UI always acts as a unified panel, one which supports
permanently closing the symboltree or calltree if desired.

initial cleaning of the in-memory trees take place in this pr but more
cleaning, such as on tab delete, will take place in a subsequent pr.

Signed-off-by: ldelossa <louis.delos@gmail.com>