ui: split jumps and "x" mapping
this commit adds the ability to jump to symbols via a split or new tab.
this is handy when encountering a symbol you'd like to explore but want
to do so in another window or tab.

additionally we add the ability press the "x" key in the calltree panel
to permanently close a particular tree.

Signed-off-by: ldelossa <louis.delos@gmail.com>