useability: make filetree jumps use the last window
this commit ensures jumping (opening) to a file from the filetree will
do so in the last source file window focused before entering the
filetree.

this is done by continually updating "invoking_filetree_win" just before
setting the current window to the filetree itself.

it short circuits if the previous focused window is the filetree itself,
ensuring the jumped-to file will not open in the filetree window.

Signed-off-by: ldelossa <louis.delos@gmail.com>