feat,popouts: this commit adds the popout window feat
this commit introduces a feature where panel components can be "popped
out" of the panel into a float on the lower right-hand side of the
editor.

a popout can occur both when the panel is closed and open. if a popout
is triggered when the panel is open the window is removed from the panel
and placed back into the panel when the "LTClosePanelPopOut" command is
issued.

if the panel is closed, it is transparently opened to generate all
component buffers and windows and the requested component window is made
into a float. the other windows are then closed.

a user could opt to use only popups and if they desire, all components
with active state will always be available when toggling the panel
regardless.