lib_lsp: method to convert SymbolInformation to DocumentSymbol
This commit adds a useful method to convert a SymbolInformation
structure to a DocumentSymbol structure.

Both structures are defined by the LSP spec and either one may be
returned by an LSP on a "documentSymbol" request.

This method helps users of the lib only deal with one data model when
working with "documentSymbol" requests.

Signed-off-by: ldelossa <louis.delos@gmail.com>